pyproject.toml
[project]
name = "tedawards"
version = "0.1.0"
description = "TED Awards scraper for EU procurement contract awards"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "lxml",
    "psycopg2-binary",
    "python-dotenv",
    "click",
    "pydantic"
]

[project.scripts]
tedawards = "tedawards.main:cli"

[tool.uv]
package = true

[tool.setuptools]
packages = ["tedawards"]
